requirements.txt
gspread
mysqlclient
oauth2client
PyYAML >= 3.12
